기흔:설문조사결과 db 테이블 insert 성공

+++ backend/src/main/java/com/jibangyoung/domain/survey/entity/SurveyAnswer.java
@@ -43,7 +43,7 @@ public class SurveyAnswer {
     private String answerText;
     
     @Column(name = "answer_weight", nullable = false)
-    private Float answerWeight;
+    private Integer answerWeight;
     
     @CreationTimestamp
     @Column(name = "submitted_at", updatable = false)
